change to float32

+++ metric/MP_LPIPS/src/models/dift_sd.py
@@ -62,7 +62,7 @@ class MyUNet2DConditionModel(UNet2DConditionModel):
             # This would be a good case for the `match` statement (Python 3.10+)
             is_mps = sample.device.type == "mps"
             if isinstance(timestep, float):
-                dtype = torch.float32 if is_mps else torch.float64
+                dtype = torch.float16 if is_mps else torch.float64
             else:
                 dtype = torch.int32 if is_mps else torch.int64
             timesteps = torch.tensor([timesteps], dtype=dtype, device=sample.device)
